Initial commit of Musical Chipmunks
So I had this problem where I wanted a network monitoring system;
but the existing systems were just not appropriate. For example
nagios or Shinken. Specifically Shinken talks about monitoring
several thousands of computers. I've tried several times to get
nagios working in various situations, but it always is way more
headache than it is worth. I needed something simplier, something
designed not for large systems but for home tinkerers.

So this is a system designed to be simple to use and simple to
setup when you only have a handful of devices you want monitored.
This initial commit has a fair bit of the structure laid out. Now
I just need to start testing and fiddling with it to get the
various checks working. One of the main tricks I'm going to depend
on is the lack of any agent that needs to be installed on the
remote system. Keep it simple for the user. Makes it more
complicated to program, but soooo simple to use.